+++ examples/test/dynarray/test_dynarray_field_badtypedconst.pas
@@ -1,0 +1,9 @@
+program TestDynArrayFieldBadTypedConst;
+type
+    TBox = record
+        data: array of integer;
+    end;
+const
+    Bad: TBox = (data: nil);
+begin
+end.